Introduce factory to create model elements and objects
