+++ subprojects/core-model/src/test/groovy/dev/nokee/model/internal/core/ModelTestUtils.java
@@ -17,6 +17,7 @@ package dev.nokee.model.internal.core;
 
 import com.google.common.collect.ImmutableList;
 import dev.nokee.model.DomainObjectProvider;
+import dev.nokee.model.internal.ModelElementFactory;
 import dev.nokee.model.internal.registry.ModelConfigurer;
 import dev.nokee.model.internal.registry.ModelLookup;
 import dev.nokee.model.internal.registry.ModelRegistry;
@@ -42,6 +43,7 @@ public final class ModelTestUtils {
 	private static final Consumer<ModelNode.Builder> DO_NOTHING = builder -> {};
 	private static final String DEFAULT_NODE_NAME = "test";
 	private static final ModelNode ROOT = rootNode();
+	private static final ModelElementFactory FACTORY = new ModelElementFactory();
 	private ModelTestUtils() {}
 
 	public static ModelProjection projectionOf(Class<?> projectionType) {
@@ -221,7 +223,7 @@ public final class ModelTestUtils {
 				val childNode = childNode(nodeProvider.getValue(), path.getName(), registration.getActions(), builder -> {});
 				registration.getComponents().forEach(childNode::addComponent);
 				children.put(path, childNode);
-				return DefaultModelElement.of(childNode);
+				return FACTORY.createElement(childNode);
 			}
 
 			private Stream<ModelPath> findModelPath(Object component) {
